feat: after user logins in successfully from google plus, app validates the login, and redirects to restaurant main page
app sends a state token to user. Then it retrieves the google authentication code from client and attempts to exchange this auth code with google for access token. if there is an error exchaning codes, the user is prompted. the code was succesfully traded, it checks to see if access token is valid by submitting a git request to google with the access token. if it is not valid, the user is prompted. if the access token is valid, check to see if the token user id matches the given user id. if it does not match, the user is prompted. if it does match, check to see if the token issue to id matches the app's id. if it does not match, the user is prompted. if the user is already logged in, show that they are already logged in. if they are not logged in, they are now logged in. their name, picture and email is retrieved from the google plus api, and it flashes for a short period of time. then they are redirected to /restaurant