feat: prot4ct restaurant menu item update to owner
if user is not logged in and they try to access the restaurant page, they will only see
the public restaurant page. logged in members will see the page with 'add restaurant'
button. if user tries to view other restaurants, they will not be able to see the
edit, delete, new restaurant button. even if they try to directly visit the URL
for edit, delete, or new restaurant for restaurants other than their own, they will
get a popup telling warning them that is not their restaurant and then get redirected
to the showrestaurant page